feat #48: OnboardingPage 임시 구현
